Use null object pattern for React Context objects.
This converts the remanining "Context" files to TypeScript and uses
no-op values for the initial state of contexts, where possible, instead
of making their values possibly null. This allows all the components
that use these contexts to use them without checking if they're null
first, which was annoying boilerplate to cover a case that doesn't
really happen in practice.