refactor: rename dim to dim-core, use dim as a binary crate (#573)
* refactor: rename dim to dim-core, use dim as a binary crate

* refactor: further organise dim-web route logic

* refactor: replace structopt for clapt+derive and remove unused deps

* Update node version to 16 in GitHub Actions workflow

The commit updates the node version from 14 to 16 in the GitHub Actions workflow file. This change ensures that the correct version of Node.js is used during the workflow execution.

* run prettier

* Remove download UI artifacts step in rust.yml

The commit removes the download UI artifacts step in the rust.yml file. This step was commented out as it is no longer needed for the workflow.